Rename some components
