+++ website/src/components/home/features.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+const Highlight = ({title, description}) => {
+	return <article>
+		<header>{title}</header>
+		<div>{description}</div>
+	</article>
+}
+
+const Highlights = () => {
+<div className="max-w-70 mx-auto flex flex-row">
+		<Highlight title="Swift manifest" description="Define projects using a simple Swift DSL inside Xcode"/>
+		<Highlight title="Project description helpers" description="Create your own abstrations to make your projects consistent"/>
+			<Highlight title="Scaffold" description="Automate feature creation by generating a target pre-configured with everything you need"/>
+	</div>
+}
+
+const Features = () => {
+	return <div className="bg-gray-900 text-white">
+		
+	</div>
+}
+
+export default Features;
