+++ website/src/components/old-layout.jsx
@@ -5,19 +5,19 @@ import React from 'react'
 import { Styled } from 'theme-ui'
 import GlobalStyle from './global-style'
 import Header from './header'
-import Footer from './footer'
+import OldFooter from './old-footer'
 import CookieBanner from './cookie-banner'
 
-const Layout = ({ children, menuOpen, setMenuOpen, menuRef }) => {
+const OldLayout = ({ children, menuOpen, setMenuOpen, menuRef }) => {
   return (
     <>
       <GlobalStyle />
       <Header menuOpen={menuOpen} setMenuOpen={setMenuOpen} menuRef={menuRef} />
       <main sx={{ mb: [3, 6] }}>{children}</main>
       <CookieBanner />
-      <Footer />
+      <OldFooter />
     </>
   )
 }
 
-export default Layout
+export default OldLayout
