+++ website/src/pages/terms.jsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx, Styled } from 'theme-ui'
 import { useStaticQuery, graphql } from 'gatsby'
-import Layout from '../components/layout'
+import OldLayout from '../components/old-layout'
 import Main from '../components/main'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
 import { GatsbySeo } from 'gatsby-plugin-next-seo'
@@ -18,7 +18,7 @@ export default () => {
     }
   `)
   return (
-    <Layout>
+    <OldLayout>
       <SEO title="Terms of Service" />
       <GatsbySeo
         title="Terms of Service"
@@ -28,6 +28,6 @@ export default () => {
         <Styled.h1>Terms of Service</Styled.h1>
         <MDXRenderer>{markdownBody}</MDXRenderer>
       </Main>
-    </Layout>
+    </OldLayout>
   )
 }
