+++ website/src/templates/apps-at-scale.jsx
@@ -2,7 +2,7 @@
 import { jsx, Styled, useThemeUI } from 'theme-ui'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
 
-import Layout from '../components/layout'
+import OldLayout from '../components/old-layout'
 import { graphql } from 'gatsby'
 import moment from 'moment'
 import Main from '../components/main'
@@ -109,7 +109,7 @@ const Page = ({
     },
   ]
   return (
-    <Layout>
+    <OldLayout>
       <BreadcrumbJsonLd itemListElements={breadcrumb} />
       <NewsArticleJsonLd
         url={urljoin(siteUrl, mdx.fields.slug)}
@@ -179,7 +179,7 @@ const Page = ({
           title={mdx.frontmatter.title}
         />
       </Main>
-    </Layout>
+    </OldLayout>
   )
 }
